+++ CHANGELOG.md
@@ -10,5 +10,6 @@ This CHANGELOG follows the format located [here](https://github.com/sensu-plugin
 ### Fixed
 - fix scrub-in-progress command
 - properly parse the list of vdevs
+- fix checking all zpools (#2)
 
 [Unreleased]: https://github.com/sensu-plugins/sensu-plugins-zfs/compare/bf20f6b2538849a9263dbaa8771d649b7173d8b1...HEAD